fix: resolve CS0266 build error - Engine.TimeScale is double, not float (Issue #544)
The Windows Export CI check failed because Engine.TimeScale returns double
in Godot's C# API, but was assigned to a float variable in
TriggerReplayPenultimateEffect(). Changed to double to match the API type.

Also updated case study with CI failure analysis and logs.

+++ Scripts/Autoload/ReplayManager.cs
@@ -2138,7 +2138,7 @@ namespace GodotTopDownTemplate.Autoload
                 if (penultimateEffects.HasMethod("_start_penultimate_effect"))
                 {
                     // Save current time_scale, trigger effect, restore time_scale
-                    float savedTimeScale = Engine.TimeScale;
+                    double savedTimeScale = Engine.TimeScale;
                     penultimateEffects.Call("_start_penultimate_effect");
                     // Restore time_scale immediately since we don't want slowdown during replay
                     Engine.TimeScale = savedTimeScale;
